requirements.txt
fastapi
uvicorn
python-dotenv
openai
requests
cryptography
sqlalchemy>=1.4,<2.0
psycopg2-binary
pydantic
python-telegram-bot==20.*
google-api-python-client
google-auth
google-auth-oauthlib
numpy<2.0
chromadb
langchain-openai
python-amazon-sp-api
redis
nest_asyncio
boto3
aiohttp
beautifulsoup4
requests_html
lxml[html_clean]